+++ back/engines/commercial/analytics/spec/services/analytics/query_runner_service_spec.rb
@@ -2,6 +2,7 @@
 
 require 'rails_helper'
 require 'query'
+require 'uri'
 
 describe Analytics::QueryRunnerService do
   describe 'run' do
@@ -16,7 +17,8 @@ describe Analytics::QueryRunnerService do
       query = Analytics::Query.new(query_param)
 
       runner = described_class.new
-      results = runner.run(query)
+      results, * = runner.run(query)
+
       expect(results).to match_array(ideas.map { |idea| { 'id' => idea.id } })
     end
 
@@ -36,7 +38,7 @@ describe Analytics::QueryRunnerService do
       query = Analytics::Query.new(query_param)
 
       runner = described_class.new
-      results = runner.run(query)
+      results, * = runner.run(query)
 
       expected_result = [
         { 'type.name' => 'initiative', 'sum_votes_count' => 1 },
@@ -63,7 +65,7 @@ describe Analytics::QueryRunnerService do
       query = Analytics::Query.new(query_param)
 
       runner = described_class.new
-      results = runner.run(query)
+      results, * = runner.run(query)
       expect(results).to eq([{ 'count' => 1 }])
     end
 
@@ -74,17 +76,76 @@ describe Analytics::QueryRunnerService do
       query_param = ActionController::Parameters.new(
         fact: 'post',
         fields: 'id',
-        sort: { id: 'ASC' },
-        limit: 3
+        sort: { id: 'ASC' }
       )
       query = Analytics::Query.new(query_param)
 
       runner = described_class.new
-      results = runner.run(query)
+      results, * = runner.run(query)
       posts = (ideas + initiatives)
         .sort_by { |p| p[:id] }
-        .map { |p| { 'id' => p.id } }[0, 3]
+        .map { |p| { 'id' => p.id } }
       expect(results).to eq(posts)
+    end
+
+    context 'result limiting and paging' do
+      before_all do
+        create_list(:idea, 10)
+      end
+
+      def build_query(size = 1, number = 1)
+        {
+          fact: 'post',
+          fields: 'id',
+          sort: { id: 'ASC' },
+          page: { size: size, number: number }
+        }
+      end
+
+      it 'returns the first page of 2 items from 10 posts' do
+        query_param = ActionController::Parameters.new(build_query(2, 1))
+        runner = described_class.new
+        results, pagination = runner.run(Analytics::Query.new(query_param))
+
+        expect(results.length).to eq(2)
+        expect(results.first['id']).to eq(Idea.order(:id).first.id)
+
+        expect(URI.decode_www_form(pagination[:self]).to_h['page[number]']).to eq('1')
+        expect(URI.decode_www_form(pagination[:first]).to_h['page[number]']).to eq('1')
+        expect(pagination[:prev]).to be_nil
+        expect(URI.decode_www_form(pagination[:next]).to_h['page[number]']).to eq('2')
+        expect(URI.decode_www_form(pagination[:last]).to_h['page[number]']).to eq('5')
+      end
+
+      it 'returns the second page of 3 items from 10 posts' do
+        query_param = ActionController::Parameters.new(build_query(3, 2))
+        runner = described_class.new
+        results, pagination = runner.run(Analytics::Query.new(query_param))
+
+        expect(results.length).to eq(3)
+        expect(results.first['id']).to eq(Idea.order(:id).fourth.id)
+        expect(results.second['id']).to eq(Idea.order(:id).fifth.id)
+
+        expect(URI.decode_www_form(pagination[:self]).to_h['page[number]']).to eq('2')
+        expect(URI.decode_www_form(pagination[:first]).to_h['page[number]']).to eq('1')
+        expect(URI.decode_www_form(pagination[:prev]).to_h['page[number]']).to eq('1')
+        expect(URI.decode_www_form(pagination[:next]).to_h['page[number]']).to eq('3')
+        expect(URI.decode_www_form(pagination[:last]).to_h['page[number]']).to eq('4')
+      end
+
+      it 'returns nothing if the page number is too high for the size' do
+        query_param = ActionController::Parameters.new(build_query(5, 3))
+        runner = described_class.new
+        results, pagination = runner.run(Analytics::Query.new(query_param))
+
+        expect(results).to eq([])
+
+        expect(URI.decode_www_form(pagination[:self]).to_h['page[number]']).to eq('3')
+        expect(URI.decode_www_form(pagination[:first]).to_h['page[number]']).to eq('1')
+        expect(URI.decode_www_form(pagination[:prev]).to_h['page[number]']).to eq('2')
+        expect(pagination[:next]).to be_nil
+        expect(URI.decode_www_form(pagination[:last]).to_h['page[number]']).to eq('2')
+      end
     end
   end
 end