+++ back/spec/lib/participation_method/native_survey_spec.rb
@@ -102,6 +102,12 @@ RSpec.describe ParticipationMethod::NativeSurvey do
     end
   end
 
+  describe '#sign_in_required_for_posting?' do
+    it 'returns false' do
+      expect(participation_method.sign_in_required_for_posting?).to be false
+    end
+  end
+
   describe '#extra_fields_category_translation_key' do
     it 'returns nil' do
       expect(participation_method.extra_fields_category_translation_key).to be_nil