+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
+## 2022-10-25
+
 ### Added
 
 - [CL-1811] New Oostende verification method
+- [CL-1558] Native surveys
 
 ## 2022-10-20